+++ libfreeipmi/include/freeipmi/driver/ipmi-kcs-driver.h
@@ -54,17 +54,17 @@ int ipmi_kcs_ctx_errnum (ipmi_kcs_ctx_t ctx);
 char *ipmi_kcs_ctx_strerror (int errnum);
 char *ipmi_kcs_ctx_errormsg (ipmi_kcs_ctx_t ctx);
 
-int8_t ipmi_kcs_ctx_get_driver_address (ipmi_kcs_ctx_t ctx, uint16_t *bmc_iobase_address);
-int8_t ipmi_kcs_ctx_get_register_spacing (ipmi_kcs_ctx_t ctx, uint8_t *register_spacing);
-int8_t ipmi_kcs_ctx_get_poll_interval (ipmi_kcs_ctx_t ctx, uint8_t *poll_interval);
-int8_t ipmi_kcs_ctx_get_flags (ipmi_kcs_ctx_t ctx, unsigned int *flags);
+int ipmi_kcs_ctx_get_driver_address (ipmi_kcs_ctx_t ctx, uint16_t *bmc_iobase_address);
+int ipmi_kcs_ctx_get_register_spacing (ipmi_kcs_ctx_t ctx, uint8_t *register_spacing);
+int ipmi_kcs_ctx_get_poll_interval (ipmi_kcs_ctx_t ctx, uint8_t *poll_interval);
+int ipmi_kcs_ctx_get_flags (ipmi_kcs_ctx_t ctx, unsigned int *flags);
 
-int8_t ipmi_kcs_ctx_set_driver_address (ipmi_kcs_ctx_t ctx, uint16_t bmc_iobase_address);
-int8_t ipmi_kcs_ctx_set_register_spacing (ipmi_kcs_ctx_t ctx, uint8_t register_spacing);
-int8_t ipmi_kcs_ctx_set_poll_interval (ipmi_kcs_ctx_t ctx, uint8_t poll_interval);
-int8_t ipmi_kcs_ctx_set_flags (ipmi_kcs_ctx_t ctx, unsigned int flags);
+int ipmi_kcs_ctx_set_driver_address (ipmi_kcs_ctx_t ctx, uint16_t bmc_iobase_address);
+int ipmi_kcs_ctx_set_register_spacing (ipmi_kcs_ctx_t ctx, uint8_t register_spacing);
+int ipmi_kcs_ctx_set_poll_interval (ipmi_kcs_ctx_t ctx, uint8_t poll_interval);
+int ipmi_kcs_ctx_set_flags (ipmi_kcs_ctx_t ctx, unsigned int flags);
 
-int8_t ipmi_kcs_ctx_io_init (ipmi_kcs_ctx_t ctx);
+int ipmi_kcs_ctx_io_init (ipmi_kcs_ctx_t ctx);
 
 int32_t ipmi_kcs_write (ipmi_kcs_ctx_t ctx,
                         uint8_t *bytes,
@@ -74,11 +74,11 @@ int32_t ipmi_kcs_read (ipmi_kcs_ctx_t ctx,
                        uint8_t* bytes,
                        unsigned int bytes_len);
 
-int8_t ipmi_kcs_cmd (ipmi_kcs_ctx_t ctx,
-                     uint8_t lun,
-                     uint8_t net_fn,
-                     fiid_obj_t obj_cmd_rq,
-                     fiid_obj_t obj_cmd_rs);
+int ipmi_kcs_cmd (ipmi_kcs_ctx_t ctx,
+                  uint8_t lun,
+                  uint8_t net_fn,
+                  fiid_obj_t obj_cmd_rq,
+                  fiid_obj_t obj_cmd_rs);
 
 #ifdef __cplusplus
 }
