+++ libfreeipmi/include/freeipmi/driver/ipmi-sunbmc-driver.h
@@ -51,19 +51,19 @@ int ipmi_sunbmc_ctx_errnum (ipmi_sunbmc_ctx_t ctx);
 char *ipmi_sunbmc_ctx_strerror (int errnum);
 char *ipmi_sunbmc_ctx_errormsg (ipmi_sunbmc_ctx_t ctx);
 
-int8_t ipmi_sunbmc_ctx_get_driver_device (ipmi_sunbmc_ctx_t ctx, char **driver_device);
-int8_t ipmi_sunbmc_ctx_get_flags (ipmi_sunbmc_ctx_t ctx, unsigned int *flags);
+int ipmi_sunbmc_ctx_get_driver_device (ipmi_sunbmc_ctx_t ctx, char **driver_device);
+int ipmi_sunbmc_ctx_get_flags (ipmi_sunbmc_ctx_t ctx, unsigned int *flags);
 
-int8_t ipmi_sunbmc_ctx_set_driver_device (ipmi_sunbmc_ctx_t ctx, const char *driver_device);
-int8_t ipmi_sunbmc_ctx_set_flags (ipmi_sunbmc_ctx_t ctx, unsigned int flags);
+int ipmi_sunbmc_ctx_set_driver_device (ipmi_sunbmc_ctx_t ctx, const char *driver_device);
+int ipmi_sunbmc_ctx_set_flags (ipmi_sunbmc_ctx_t ctx, unsigned int flags);
 
-int8_t ipmi_sunbmc_ctx_io_init (ipmi_sunbmc_ctx_t ctx);
+int ipmi_sunbmc_ctx_io_init (ipmi_sunbmc_ctx_t ctx);
 
-int8_t ipmi_sunbmc_cmd (ipmi_sunbmc_ctx_t ctx,
-                        uint8_t lun,
-                        uint8_t net_fn,
-                        fiid_obj_t obj_cmd_rq,
-                        fiid_obj_t obj_cmd_rs);
+int ipmi_sunbmc_cmd (ipmi_sunbmc_ctx_t ctx,
+                     uint8_t lun,
+                     uint8_t net_fn,
+                     fiid_obj_t obj_cmd_rq,
+                     fiid_obj_t obj_cmd_rs);
 
 #ifdef __cplusplus
 }
